set up decorator for member roles in institutions
