+++ helpers/utils.py
@@ -388,22 +388,3 @@ def discoverable_project_view(project, user):
         discoverable = False
 
     return discoverable
-
-def set_ror_id(institution):
-    import requests
-
-    url = 'https://api.ror.org/organizations'
-    query = institution.institution_name
-    params = { 'query': query }
-    response = requests.get(url, params=params)
-
-    if response.status_code == 200:
-        data = response.json()
-        if 'items' in data and len(data['items']) > 0:
-            ror_id = data['items'][0]['id']
-            institution.ror_id = ror_id
-            institution.save()
-        else:
-            print('No matching institution found.')
-    else:
-        print('Error:', response.status_code)
