+++ institutions/utils.py
@@ -1,0 +1,24 @@
+from .models import Institution
+
+def get_institution(pk):
+    return Institution.objects.select_related('institution_creator').prefetch_related('admins', 'editors', 'viewers').get(id=pk)
+
+# TODO: where is this used??
+def set_ror_id(institution):
+    import requests
+
+    url = 'https://api.ror.org/organizations'
+    query = institution.institution_name
+    params = { 'query': query }
+    response = requests.get(url, params=params)
+
+    if response.status_code == 200:
+        data = response.json()
+        if 'items' in data and len(data['items']) > 0:
+            ror_id = data['items'][0]['id']
+            institution.ror_id = ror_id
+            institution.save()
+        else:
+            print('No matching institution found.')
+    else:
+        print('Error:', response.status_code)
